ai-agent/services/api.py
from fastapi import FastAPI, HTTPException, Depends
from pydantic import BaseModel
import os
import sys
import logging
from typing import List, Optional
from dotenv import load_dotenv

# Add parent directory to path
sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))

# Load environment variables
load_dotenv()

# Import core components
# These will be implemented as needed

# Setup logging
logging.basicConfig(
    level=getattr(logging, os.getenv("LOG_LEVEL", "INFO")),
    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
)
logger = logging.getLogger("api-service")

# Initialize FastAPI app
app = FastAPI(
    title="AI Agent API",
    description="API for interacting with the AI Agent",
    version="0.1.0",
)

# Models
class ChatRequest(BaseModel):
    message: str
    context: Optional[dict] = None

class ChatResponse(BaseModel):
    response: str
    actions: Optional[List[dict]] = None
    context: Optional[dict] = None

# Routes
@app.get("/")
async def root():
    return {"status": "active", "service": "AI Agent API"}

@app.post("/chat", response_model=ChatResponse)
async def chat(request: ChatRequest):
    try:
        # This will be implemented to call the actual AI agent
        logger.info(f"Received chat request: {request.message}")
        
        # Placeholder response
        return ChatResponse(
            response="This is a placeholder response. The actual AI agent integration is not implemented yet.",
            actions=[],
            context=request.context
        )
    except Exception as e:
        logger.error(f"Error processing chat request: {str(e)}")
        raise HTTPException(status_code=500, detail=str(e))

# Run the app
if __name__ == "__main__":
    import uvicorn
    uvicorn.run(
        "api:app",
        host=os.getenv("API_HOST", "0.0.0.0"),
        port=int(os.getenv("API_PORT", 8000)),
        reload=os.getenv("DEBUG", "False").lower() == "true",
    )
