ai-agent/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pyttsx3>=2.90
SpeechRecognition>=3.10.0
requests>=2.31.0

# For RAG and vector storage
faiss-cpu>=1.7.4
langchain>=0.0.270
sentence-transformers>=2.2.2

# For API and services
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0.3

# Utilities
numpy>=1.24.0
pydantic>=2.0.0
pytest>=7.4.0
pillow>=10.0.0
